Explore our Tasks Controller...
The routes.rb GETs the request,
connects that request to tasks#index,
and TaskController's index action passes a variable
to the view containing all the tasks. Like this:
(in app/controllers/tasks_controller.rb):
def index
  @tasks = Task.all
end
The view that receives the variable with all the tasks
(the @tasks variable), is in app/views/tasks/index.html.erb
what it does with that variable?
It runs it through a loop.
Focus on this line in app/v/t/i.h.e:
  <tbody>
   <% @tasks.each do |task| %>
   <tr>
      <td><%= task.content %></td> # displays content
      <td><%= link_to 'Show', task %></td> # links to the task#show
      etc...

The next action in TaskController
is the show action (task#show).
If you read the comment, you'll notice
that it responds to the GET request,
and will GET /tasks/1.

How does rails know which task we want?
If we inspect the tasks_controller.rb,
we see on the very top of the file,
a before_action :set_task,
and on the bottom of the controller
we see that there is a private section,
with a set_task method.

This private method is where a variable is
created, just like with the index action,
to be sent to the view.

Only this time, we don't assign All the tasks,
but we find the task, with the params that
the browser has sent us.

In this case, the request get/tasks/1,
is sent to the controller, and the
controller  set_task method will search
for the task with the id: 1,
and set it to the @task variable.

As you see, the before action sets the right task for us
using the params it received, every time the show,
edit, update or destroy action is called.

It would not make sense to assign a task to the new action,
it's not even possible because the task doesn't exist yet.
So, for the :new action, we simply create a new empty task,
and assign it to the variable @task.

The :create, :update, and :delete actions look more complicated,
but they follow the same principles, the main big difference is
that they have a so called respond_to block.
This will handle what happens
if the request was handled successfully.

For instance, if we want our application to use javascript to
add the newly created task to our list, we could use the format.js,
and specify what to do.
One other thing you'll probably have al ready noticed,
is the task_params. Rails is paranoid about safety, and won't let
you pass any data from the view to the database without permission:

  # Never trust parameters from the scary internet,
  only allow the white list through.
  def task_params
    params.require(:task).permit(:content)
  end

That's why, inside of the :create and :update actions,
you'll notice the use of the task_params action.

That's all we have to know for now about controllers,
this may seem complicated, but you'll get the basics soon enough.
Just remember to use rake routes to know which verb to connect to
what action.

In the next chapter, we will explore the views.