Explore our task model, and add a user to a task...
Add a user id to tasks, configured user task associations

go to <site>/tasks

create a new task as User1
now log out, sign up as User2

go to <site>/tasks
and User2 will see User1's task

we haven't settled associations yet, so the task
doesn't belong to anyone, it belongs to itself

for the 1st time, we'll use the rails console
it will let us use rails operations inside erb

in terminal, type 'rails console'

> Task
> (call 'task.connection' to establish a connection)

ok, we'll do it like it asks, so:
> Task.connection
> (we get a connection)
> exit
> (we exit the rails console)
> clear
> (this clears the screen)
> rails console
> Task
> Task.connection
> Task
> (id: integer, content: text, created_at: datetime, updated_at:
> datetime)

What we did just now is inspect the Task class and it returned the
fields that it has. Like tasks'id, what's its contents,
when it was created_at, what datetime it was updated_at.

But our app still doesn't have any way of identifying which user
a task belongs to.

So we first have to create a new migration, add sth to the DB for
the Task table, that says "this task belongs to a specific user".

So let's exit RC and generate a migration:
----------------------------------------------------------
rails generate migration AddUser_idToTasks user_id:integer
----------------------------------------------------------

this will create a new migration in app/db/migrate, as follows:

class AdduserIdToTasks < ActiveRecord::Migration
  def change
      add_column :tasks, :user_id, :integer
  end
end

Looking at the code, this is what is does:
add a column to our Tasks table
the column name is 'user_id'
the value that can go into 'user_id' is integer

How do we add this to the DB?
We run:
---------------
rake db:migrate
---------------

When done, run the rails console:
> rails console
> Task.connection

Let's inspect our Task class and see if we have anything new:
> Task
> Task(id: integer, content: text, created_at: datetime
> updated_at: datetime, user_id: integer)
> exit

Now let's look at /models.
This is where we put any associations.

in models/user.rb, we'll make the following association:
    a user can have many tasks
    and
    a task belongs to a user

under the devise code in user.rb, for class User we say
that he/she:

has_many :tasks, dependent: :destroy

the meaning of dependent: :destroy is that if a user
is deleted, its tasks are deleted, so that there are no
'hanging tasks' that don't belong to anybody.

now in models/task.rb, for class Task, we say
that it:
    belongs_to :user

Now all that is left is that we need to on the creation
associate the task created with a specific user:

in app/controllers/tasks_controller.rb, in our index action,
we're gonna modify this line:
@tasks = Task.all

and we'll use devise method current_user:
@tasks = current_user.tasks

We need to do the same for def create:
delete this:
@tasks = Task.new(task_params)
replace it with this:
@task = current_user.tasks.new(task_params)

Save all.
> rails server

Verify that you do not see the tasks that belong to the other user.

in tasks_controller.rb, we'll enforce something,
what happens when they try to fetch <site>/tasks when they are
not logged in.

You get "NoMethodError in TasksController#index"
Why is that?>

Well, in your controller, you just siad, in your index action
to fetch the current_user.tasks

Well, since no user is currently logged in, how do you expect to
fetch the tasks for a current_user that is not there?

Fortunately, devise has a cool method (google: devise),
github devise docs title: Controller filters and helpers

before_action :authenticate_user!

we gonna add this to our controller.
Which means, in order to run a specific action, the
user must be logged in, and since we want that for
every specific action, cuz we don't want the user to access
it while logged out, we are gonna paste this line...

before_action :authenticate_user!

... just under the text "class TasksController < Appli..."

if we try to access <site>/tasks now, without being logged in,
it will redirect us to the users login page.

commit with the above text as message