Add custom actions...
It would be easier to add a btn to each task to change its state.

We'll need to add a custom action called change.

so let's go to the controller, and create this custom action
beneath the destroy action:

def change
end

rails console
(this will help us see better what we're trying to do)

make new task
task = Task.create(content: "hello")

remember in previous commit we added the default state.
to test our task, let's run:

task

let's assign this task a user

task.user = User.first

let's save this task

task.save

let's inspect the id and the rest:

task

yes, it lists everything

let's modify the state:

task.state = "doing"

save it:

task.save

inspect it:

task

Notice there are 2 steps involved to make this happen.
The first step is setting the variable to a different string,
second is saving the task.

There is an easier way of doing this, and it's called
"update attributes", and with 1 line of code we can to it,
modify our field and save it at the same time, and that's
exactly what we'll do in our controller.

Still in rails console let's test it:

task.update_attributes(state: "done")

let's see if it works:

task

now we can see that the state is "done"

Let's now add this functionality to def change

in task controller, we need to list change (on top of the file), like this:

before_action :set_task, only: [:show, :edit, :update, :destroy, :change]

now in def change, we're going to copy the logic from the
private set_task method, which is passed a params[:id]:

def change
  @task.update_attributes(state: params[:state])

Now we wanna redirect this to the tasks page:

def change
  @task.update_attributes(state: params[:state])
  respond_to do |format|

In respond_to do |format|, what we're saying is, after the
line above is complete, set the format of the resulting file
(it can be html or js, we wanna use html cuz we wanna
redirect to index after it's updated), so:

def change
  @task.update_attributes(state: params[:state])
  respond_to do |format|
    format.html {redirct_to tasks_path, notice: "Task Updated"}
  end
end

Now we need to add a custom route to our routes file, so our
app knows where to look for when we ask to change our task.

Before we do this, it's a good idea to explain CRUD.

Operation | HTTP
Create    | POST
Read      | GET
Update    | PUT
Delete    | DELETE

We can see this working on our rake routes.

So task#create correlates to a POST.

Since our last action is a custom action, we have to tell Rails
how to handle the request.

It's just an Update operation. So it must be an HTTP PUT request.

We do this by modifying a routes file.

in app/config/routes.rb:

change resources :tasks into a block:

from...
resources :tasks
...to
resources :tasks do
  member do
    put :change
  end
end

this will tell our routes that there is a new route open to
:change custom action.

Now we need to find what our path available is.
We do this with rake routes:

(now it lists our :change custom action's route, like this:)

Prefix     |   Verb    |   URI Pattern   |   Controller#Action
change_task|    PUT    |  /tasks/:id/change(.:format) | tasks#change

Basically, the only thing it asks for is the id, which is normal.

So let's go back to app/views/tasks/index.html.erb, and make the
links work:

So for each different state :change request, we will have to
make different links. In other words, we will have to make
a link for each set of these actions.

So for change_task to "doing":
<%= link_to "Mark as Doing",
change_task_path(task, state: "doing"),
method: :put %>

for "done":
<%= link_to "Mark as Done", change_task_path(task, state: "done"),
method: :put %>

for "to do":
<%= link_to "Mark as To Do", change_task_path(task, state: "to do"),
method: :put %>

"Mark as To Do" is the text that's going to appear on the btn
change_task_path is the route converted to the link format by
adding the _path part
(task, state: "to do") the task part tells the controller which
task it is and the second part tells it which state is being
assigned.

Now we just need to add this to our view, but where should we
put each link?

  To Do       |     Doing     |     Done
Mark as Doing | Mark as To Do | Mark as Doing
              | Mark as Done  |

As a guide, we'll refer to the table above.

now in app/views/tasks/index.html.erb

from...
<th colspan="3">
...to
<th colspan="4">

in...
<tbody>
  <% @to_do.each...,
...before (above)...
</tr>
...add...
<td><%= link_to "Mark as Doing", change_task_path(task, state:
 "doing"), method: :put %></td>

 let's look at it in browser.

 We shoudl see an extra link in "To Do"

 Now add links to "Doing" and "Done":

in "Doing"
<th colspan="5">
<td><%= link_to "Mark as To Do", change_task_path(task, state:
"to_do"), method: :put %></td>
<td><%= link_to "Mark as Done", change_task_path(task, state:
"done"), method: :put %></td>

in "Done"
<th colspan="4">
<td><%= link_to "Mark as Doing", change_task_path(task, state:
 "doing"), method: :put %></td>

Save all and check.

Commit with this message.