Make tasks look better with Bootstrap... (Apply new design to old task logic)
Time to add logic to our app/view/tasks/index.html.erb

go to <site>/tasks

we're not gonna use tables, so:
delete all the content from app/views/tasks/index.html.erb

... except the <br> and the one line below it

...back to app/views/pages/test.html.erb, copy

... in app/views/tasks/index.html.erb,

add .row

inside the row, add:

.col-md-4>h1.text-center*3(to do, doing, done)

in app/views/tasks/_task.html.erb comment out everything using the <!-- html comment, then under the comment,

USE The panel (comes
from app/views/pages/test.html.erb) and
paste the panel at the bottom of the file _task file.

html comment out everything above the pasted panel

in a/v/t/i.h.erb, add the renders in between the md-4 divs, just before each </div>:

<%= render @to_do %>
<%= render @doing %>
<%= render @done %>

but this only provides the panels without the specific info

so we're gonna use the helper "time_ago_in_words"

the helper goes before, and the method goes after

<div class="panel panel-info">
  <div class="panel-heading">created <%= time_ago_in_words(task.created_at) %> ago</div>
  <div class="panel-body"><%= task.content %></div>
  <div class="panel-footer text-center"></div>

let's copy the html commented out conditions and paste them above the .panel-footer line. Remove the <td>'s as we are not using tables any more.

All the links look good, but it is better to replace the written links with .fa.fa-arrow-right.

So we have to alter the conditions, by adding the else-if condition.

Like this:

<% if task.state == "doing" %>
<%= link_to "<i class='fa fa-arrow-left'></i>"

remember you need to escape it!
let's try it again:
<% if task.state == "doing" %>
<%= link_to "<i class='fa fa-arrow-left'></i>".html_safe, change_task_path(task, state: "to_do"), method: :put %>
<%= link_to "<i class='fa fa-arrow-right'></i>".html_safe, change_task_path(task, state: "done"), method: :put %>
<% elsif task.state == "done" %>
<%= link_to "<i class='fa fa-arrow-left'></i>".html_safe, change_task_path(task, state: "doing"), method: :put %>
<% else %>
<%= link_to "<i class='fa fa-arrow-right'</i>".html_sfe, change_task_path(task, state: "doing"), method: :put %>

now let's make the arrows into buttons, we'll add the button code at the end of each line, like this:
<%= link_to "<i class='fa fa-arrow-left'></i>".html_safe, change_task_path(task, state: "to_do"), method: :put, class: "btn btn-info btn-md btn-block" %>

So let's add these to all of the links.

however, the buttons are right now stacking, and we want them side by side.

So, we need to group these buttons together, so we'll use button group.

just under the <% if...
insert another line:
<div class="btn-group btn-group-justified">

finally, let's add teh links to edit the created at to be a link:

rake routes,
we're looking for edit_task_path, so let's do it linke this:
(we're making the entire content of panel-heading into a button):

<div class="panel-heading">
<%= link_to edit_task_path(task) do %>
created <%= time_ago_in_words(task.created_at) %> ago
<% end %>
</div>

(to make everything a link, we added the do at the end of the line under the <div class="panel-heading"). Adding a 'do' will make it behave like a block.

we don't need the commented out html, cuz we have used it already and applied all the neeed logic.

git commit