Finish ui ux touches, add ability to destroy tasks, style edit task and new task form, hide jumbotron when logged in, ...
we need to be able to destroy tasks from time to time, so they don't clog up.

rake routes

tasks#destroy is a DELETE method, and the link is a bit different

in app/views/tasks/_task.html.erb

after the <% end %> in panel heading, add:

<%= link_to "x", task_path(task), method: :delete, data: {confirm: "Are you sure?"} %>

this data: part is kicking a js function that makes you confirm your action.

It works, so let's style it a bit:
<%= link_to "x", task_path(task), method: :delete, data: {confirm: "Are you sure?"}, class: "pull-right" %>

Now in tasks/new, we don't need the state any more, and we do need to style it up a bit. It's the same as with tasks/edit.

In app/views/tasks/_form, remove the unnecessary divs, that were created by the rails scaffold.

Remove the state as we have the default. Save, refresh.

in our simple_form_for, we're gonna give it an extra class:
<%= simple_form_for(@task, html: {class: "form-signin"}) do |f| %>

we don't need a lable any more:

<%= f.input :content, label: false, required: true, autofocus: true, input_html: {class: "form-control"}, placeholder: "Content" %>

save and refresh

Finally, center "Edit task" and restyle the button:

<%= f.button :submit, class: "btn btn-lg btn-primary btn-block" %>

to center "Edit task" and "New task", we have to add .text-center to new.html.erb and edit.html.erb in app/views/tasks.

Finally, we'll create sth in the pages_controller.rb in app/controllers, in def home:

def home
  if current_user
    redirect_to tasks_path
  end
end

so if the user is logged in he will never see the sign up jumbotron

commit